+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to the "shebang-language-associator" extension will be documented in this file.
 
 ## [1.3.0]
-- The extension is activated upon saving files. In previous versions, the extension was only activated upon installation and when opening files.
+- Add new option `shebang.associateOnSave` (default `true`) to also check files for language associations on save. Previous versions of the extension only ran on file open.
 
 ## [1.2.0]
 - Add support for zsh and scripts run by the default shell (such as golang scripts).